requirements.txt
paho-mqtt==1.6.1
pandas==1.5.3
#git+https://github.com/online-ml/river.git
river==0.12.0
scipy==1.10.0
streamz==0.6.4
streamz_pulsar==0.1.0.post1
confluent-kafka==2.0.2
matplotlib==3.7.0
plotly==5.13.0
cryptography==41.0.1
human_security==1.0